More refractoring
- versioning uses a double
- edited serviceattribute
- separated StackWrapper in many shorter files
