+++ StackInjector/Settings/DependencyGraphActions.cs
